+++ admin/.env.example
@@ -1,2 +1,5 @@
-NEXT_PUBLIC_APP_URL=
-NEXT_PUBLIC_API_BASE_URL=
+NEXT_PUBLIC_API_BASE_URL=""
+NEXT_PUBLIC_ADMIN_BASE_URL=""
+NEXT_PUBLIC_SPACE_BASE_URL=""
+NEXT_PUBLIC_WEB_BASE_URL=""
+NEXT_PUBLIC_SPACE_BASE_PATH="/spaces"
