+++ admin/Dockerfile.admin
@@ -1,3 +1,6 @@
+# *****************************************************************************
+# STAGE 1: Build the project
+# *****************************************************************************
 FROM node:18-alpine AS builder
 RUN apk add --no-cache libc6-compat
 WORKDIR /app
@@ -7,6 +10,9 @@ COPY . .
 
 RUN turbo prune --scope=admin --docker
 
+# *****************************************************************************
+# STAGE 2: Install dependencies & build the project
+# *****************************************************************************
 FROM node:18-alpine AS installer
 
 RUN apk add --no-cache libc6-compat
@@ -21,13 +27,25 @@ COPY --from=builder /app/out/full/ .
 COPY turbo.json turbo.json
 
 ARG NEXT_PUBLIC_API_BASE_URL=""
-ARG NEXT_PUBLIC_ADMIN_BASE_PATH="/god-mode"
-
 ENV NEXT_PUBLIC_API_BASE_URL=$NEXT_PUBLIC_API_BASE_URL
+
+ARG NEXT_PUBLIC_WEB_BASE_URL=""
+ENV NEXT_PUBLIC_WEB_BASE_URL=$NEXT_PUBLIC_WEB_BASE_URL
+
+ARG NEXT_PUBLIC_SPACE_BASE_URL=""
+ENV NEXT_PUBLIC_SPACE_BASE_URL=$NEXT_PUBLIC_SPACE_BASE_URL
+
+ARG NEXT_PUBLIC_ADMIN_BASE_PATH="/god-mode"
 ENV NEXT_PUBLIC_ADMIN_BASE_PATH=$NEXT_PUBLIC_ADMIN_BASE_PATH
+
+ENV NEXT_TELEMETRY_DISABLED 1
+ENV TURBO_TELEMETRY_DISABLED 1
 
 RUN yarn turbo run build --filter=admin
 
+# *****************************************************************************
+# STAGE 3: Copy the project and start it
+# *****************************************************************************
 FROM node:18-alpine AS runner
 WORKDIR /app
 
@@ -38,11 +56,16 @@ COPY --from=installer /app/admin/.next/standalone ./
 COPY --from=installer /app/admin/.next/static ./admin/.next/static
 COPY --from=installer /app/admin/public ./admin/public
 
-
 ARG NEXT_PUBLIC_API_BASE_URL=""
-ARG NEXT_PUBLIC_ADMIN_BASE_PATH="/god-mode"
-
 ENV NEXT_PUBLIC_API_BASE_URL=$NEXT_PUBLIC_API_BASE_URL
+
+ARG NEXT_PUBLIC_WEB_BASE_URL=""
+ENV NEXT_PUBLIC_WEB_BASE_URL=$NEXT_PUBLIC_WEB_BASE_URL
+
+ARG NEXT_PUBLIC_SPACE_BASE_URL=""
+ENV NEXT_PUBLIC_SPACE_BASE_URL=$NEXT_PUBLIC_SPACE_BASE_URL
+
+ARG NEXT_PUBLIC_ADMIN_BASE_PATH="/god-mode"
 ENV NEXT_PUBLIC_ADMIN_BASE_PATH=$NEXT_PUBLIC_ADMIN_BASE_PATH
 
 ENV NEXT_TELEMETRY_DISABLED 1
