+++ admin/components/admin-sidebar/help-section.tsx
@@ -1,12 +1,13 @@
 "use client";
 
 import { FC, useState, useRef } from "react";
-import { Transition } from "@headlessui/react";
 import Link from "next/link";
+import { observer } from "mobx-react-lite";
+import { Transition } from "@headlessui/react";
 import { ExternalLink, FileText, HelpCircle, MoveLeft } from "lucide-react";
 import { DiscordIcon, GithubIcon, Tooltip } from "@plane/ui";
 // hooks
-import { useTheme } from "@/hooks";
+import { useInstance, useTheme } from "@/hooks";
 // assets
 import packageJson from "package.json";
 
@@ -28,7 +29,9 @@ const helpOptions = [
   },
 ];
 
-export const HelpSection: FC = () => {
+export const HelpSection: FC = observer(() => {
+  // hooks
+  const { instance } = useInstance();
   // states
   const [isNeedHelpOpen, setIsNeedHelpOpen] = useState(false);
   // store
@@ -36,7 +39,7 @@ export const HelpSection: FC = () => {
   // refs
   const helpOptionsRef = useRef<HTMLDivElement | null>(null);
 
-  const redirectionLink = `${process.env.NEXT_PUBLIC_APP_URL ? `${process.env.NEXT_PUBLIC_APP_URL}/create-workspace` : `/god-mode/`}`;
+  const redirectionLink = `${instance?.config?.app_base_url ? `${instance?.config?.app_base_url}/create-workspace` : `/god-mode/`}`;
 
   return (
     <div
@@ -128,4 +131,4 @@ export const HelpSection: FC = () => {
       </div>
     </div>
   );
-};
+});
