+++ apiserver/.env.example
@@ -44,3 +44,8 @@ WEB_URL="http://localhost"
 
 # Gunicorn Workers
 GUNICORN_WORKERS=2
+
+# Base URLs
+ADMIN_BASE_URL=
+SPACE_BASE_URL=
+APP_BASE_URL=
