+++ apiserver/plane/api/views/base.py
@@ -1,6 +1,4 @@
 # Python imports
-from urllib.parse import urlparse
-
 import zoneinfo
 
 # Django imports
