+++ apiserver/plane/authentication/urls.py
@@ -7,6 +7,7 @@ from .views import (
     ForgotPasswordEndpoint,
     SetUserPasswordEndpoint,
     ResetPasswordEndpoint,
+    ChangePasswordEndpoint,
     # App
     GitHubCallbackEndpoint,
     GitHubOauthInitiateEndpoint,
@@ -18,6 +19,8 @@ from .views import (
     SignInAuthEndpoint,
     SignOutAuthEndpoint,
     SignUpAuthEndpoint,
+    ForgotPasswordSpaceEndpoint,
+    ResetPasswordSpaceEndpoint,
     # Space
     EmailCheckEndpoint,
     GitHubCallbackSpaceEndpoint,
@@ -174,6 +177,21 @@ urlpatterns = [
     path(
         "reset-password/<uidb64>/<token>/",
         ResetPasswordEndpoint.as_view(),
+        name="forgot-password",
+    ),
+    path(
+        "spaces/forgot-password/",
+        ForgotPasswordSpaceEndpoint.as_view(),
+        name="forgot-password",
+    ),
+    path(
+        "spaces/reset-password/<uidb64>/<token>/",
+        ResetPasswordSpaceEndpoint.as_view(),
+        name="forgot-password",
+    ),
+    path(
+        "change-password/",
+        ChangePasswordEndpoint.as_view(),
         name="forgot-password",
     ),
     path(
