+++ apiserver/plane/license/api/views/instance.py
@@ -2,6 +2,7 @@
 import os
 
 # Django imports
+from django.conf import settings
 
 # Third party imports
 from rest_framework import status
@@ -148,9 +149,13 @@ class InstanceEndpoint(BaseAPIView):
         )
 
         # is smtp configured
-        data["is_smtp_configured"] = (
-            bool(EMAIL_HOST)
-        )
+        data["is_smtp_configured"] = bool(EMAIL_HOST)
+
+        # Base URL
+        data["admin_base_url"] = settings.ADMIN_BASE_URL
+        data["space_base_url"] = settings.SPACE_BASE_URL
+        data["app_base_url"] = settings.APP_BASE_URL
+
         instance_data = serializer.data
         instance_data["workspaces_exist"] = Workspace.objects.count() > 1
