+++ space/.env.example
@@ -1,2 +1,3 @@
-NEXT_PUBLIC_APP_URL=
-NEXT_PUBLIC_API_BASE_URL=
+NEXT_PUBLIC_API_BASE_URL=""
+NEXT_PUBLIC_WEB_BASE_URL=""
+NEXT_PUBLIC_SPACE_BASE_PATH="/spaces"
