+++ turbo.json
@@ -3,9 +3,11 @@
   "globalEnv": [
     "NODE_ENV",
     "NEXT_PUBLIC_API_BASE_URL",
-    "NEXT_PUBLIC_APP_URL",
-    "NEXT_PUBLIC_DEPLOY_URL",
-    "NEXT_PUBLIC_GOD_MODE_URL",
+    "NEXT_PUBLIC_ADMIN_BASE_URL",
+    "NEXT_PUBLIC_ADMIN_BASE_PATH",
+    "NEXT_PUBLIC_SPACE_BASE_URL",
+    "NEXT_PUBLIC_SPACE_BASE_PATH",
+    "NEXT_PUBLIC_WEB_BASE_URL",
     "NEXT_PUBLIC_SENTRY_DSN",
     "NEXT_PUBLIC_SENTRY_ENVIRONMENT",
     "NEXT_PUBLIC_ENABLE_SENTRY",
@@ -18,8 +20,7 @@
     "NEXT_PUBLIC_POSTHOG_KEY",
     "NEXT_PUBLIC_POSTHOG_HOST",
     "NEXT_PUBLIC_POSTHOG_DEBUG",
-    "SENTRY_AUTH_TOKEN",
-    "NEXT_PUBLIC_SPACE_BASE_PATH"
+    "SENTRY_AUTH_TOKEN"
   ],
   "pipeline": {
     "build": {
