+++ web/.env.example
@@ -1,2 +1,7 @@
-# Public boards deploy URL
-NEXT_PUBLIC_DEPLOY_URL="http://localhost/spaces"
+NEXT_PUBLIC_API_BASE_URL=""
+
+NEXT_PUBLIC_ADMIN_BASE_URL=""
+NEXT_PUBLIC_ADMIN_BASE_PATH="/god-mode"
+
+NEXT_PUBLIC_SPACE_BASE_URL=""
+NEXT_PUBLIC_SPACE_BASE_PATH="/spaces"
